feat(dashboard): split skill detail into routed pages
